Remove pass through hooks.
